Entry Point
This is the entry point for this project.  The idea was to begin learning C++ after already having grasped concepts from programming elsewhere both in C++ and C#.  The aim of this project is to create a foundation of a 3D graphics engine that will later be developed into a functioning game.